refactor(pstat): Use int fields; cap downcasts in getpinfo()
Simplify user-visible pstat to int and guard downcasts.
- pstat.{tickets,pid,ticks}: int
- getpinfo(): clamp uint/uint64 to INT_MAX before copyout
- types.h: define INT_MAX for clamping
Notes: pstat is a user ABI; update user tools accordingly.

+++ kernel/pstat.h
@@ -6,9 +6,9 @@
 
 struct pstat {
   int inuse[NPROC];   // whether this slot of the process table is in use (1 or 0)
-  uint tickets[NPROC]; // the number of tickets this process has
-  uint pid[NPROC];     // the PID of each process
-  uint64 ticks[NPROC];   // the number of ticks each process has accumulated
+  int tickets[NPROC]; // the number of tickets this process has
+  int pid[NPROC];     // the PID of each process
+  int ticks[NPROC];   // the number of ticks each process has accumulated
 };
 
 #endif // _PSTAT_H_
